+++ test/Spec.hs
@@ -3,13 +3,15 @@ module Spec where
 import Test.Hspec
 
 import qualified CBORSpec as C
+import qualified ClientSpec as Client
 import qualified HTTPSpec as H
 import qualified MiscSpec as M
 import qualified SemanticSpec as S
 
 spec :: Spec
-spec = do
-    parallel $ describe "HTTP" H.spec
-    parallel $ describe "Misc" M.spec
-    parallel $ describe "Semantic" S.spec
-    parallel $ describe "CBOR" C.spec
+spec = parallel $ do
+    describe "HTTP" H.spec
+    describe "Misc" M.spec
+    describe "Semantic" S.spec
+    describe "CBOR" C.spec
+    describe "Client" Client.spec